# cmd/api-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: Tickstream API
  version: 1.0.0
  description: |
    Live market metrics, alert history, feed health and backtest reports.
    Times are nanoseconds since the Unix epoch (UTC). Prices, VWAPs and
    volumes are decimal strings so no precision is lost.

    Live updates are pushed over a websocket at `GET /ws?symbols=BTC-USD,ETH-USD`
    (omit `symbols` for all). Each message is `{"type": "metrics" | "alert" |
    "health", "data": <Metrics | Alert | FeedHealth>}`. Slow clients may miss
    old metrics messages, never alerts.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/alerts:
    get:
      summary: Alert history, newest first
      description: |
        With `symbol`, results can be paged with `cursor` (use `next_cursor`
        from the previous page). Without `symbol`, the newest alerts of all
        symbols are merged and paging is not available.
      operationId: listAlerts
      parameters:
        - name: symbol
          in: query
          schema: { type: string, example: BTC-USD }
        - name: since
          in: query
          description: Only alerts for windows ending at or after this time (ns).
          schema: { type: integer, format: int64 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertPage" }
        "400": { $ref: "#/components/responses/Error" }
...
  /api/v1/backtests/{run_id}:
    get:
      summary: A backtest report (the report.json written by the backtester)
      operationId: getBacktest
      parameters:
        - name: run_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "404": { $ref: "#/components/responses/Error" }
...
components:
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
  schemas:
...
    Alert:
      type: object
      required: [symbol, alert_id, rule_name, severity, window_end_ns, message, values]
      properties:
        symbol: { type: string }
        alert_id: { type: string }
        rule_name: { type: string }
        severity: { type: string, enum: [INFO, WARN, CRITICAL] }
        window_end_ns: { type: integer, format: int64 }
        exchange: { type: string, description: Set for per-exchange rules such as stale_feed }
        message: { type: string }
        values:
          type: object
          additionalProperties: { type: number }
    AlertPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Alert" }
        next_cursor: { type: string }
...